Create component: Add product and add return button
