+++ front_end/src/pages/Admin/CategoryDashboard/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./style.css";
-
+import AddProduct from "../../../components/AddProduct";
 import {
     Card,
     Container,
@@ -64,7 +64,10 @@ export default function CategoryDashboard(){
           text ="Danh mục sản phẩm"
           count = "6"/>
           </Row>
+        <div className="d-flex justify-content-between">
           <Link to="/admin/all-products"><h5 className="p-2">Xem tất cả {'>>'} </h5></Link>
+            <AddProduct/>
+        </div>
           <Row>  
           {type.map((item, index) => {
           return (
