+++ front_end/src/pages/Admin/OrderDetail/index.jsx
@@ -9,7 +9,7 @@ import { ToastContainer, toast } from 'react-toastify';
 
 import 'react-toastify/dist/ReactToastify.css';
 import Sidebar from "../../../components/sidebar/Sidebar";
-import { useParams} from "react-router-dom";
+import { useParams, Link} from "react-router-dom";
 import AdminNavbar from "../../../components/NavBar/Navbar";
 import moment from "moment";
 import axios from "../../../api/axios";
@@ -86,7 +86,8 @@ export default function OrderDetail(){
     <Col>
     <Row>
     <AdminNavbar 
-    title="Quản lý đơn hàng"/>
+    title="Quản lý đơn hàng"
+    subtitle= {order_id}/>
     <Card className="card-plain table-plain-bg">
     <Card.Body className="table-full-width table-responsive px-0">
       <Table>
@@ -235,6 +236,8 @@ export default function OrderDetail(){
       </div>
 
     </Row>
+    <Link to="/admin/order"><h5 className="p-2">{'<<'} Trở về </h5></Link>
+
     </Col>
       </Row>
